MLbackend/requirements.txt
fastapi
uvicorn
tensorflow
ultralytics
pillow
numpy
python-multipart
